Reorganize file directory
